broken database
First time I try updating the database and reading items from it. It doesn't seem to be working currently, as only the boilerplate code for GetTonight() from the tutorial seems to have any positive outcome when trying to read anything written to the database. ExampleInstrumentedTest also did not execute properly.