Fixed GameFragment DB reading problem with GC
The problem was in GameFragment observer block, viewModel.closeGameDialog() was infinite looping observer. Fixed Load Game screen so it properly navigates to GameFragment rather than SetupFragment if the Status is "Playing Game", to SetupFragment if the Status is "Setup", and just prints an error with any other status (because NYI).